Saving scrolled position into MovieListViewModel and SearchViewModel
