+++ movies/src/main/java/com/klewerro/moviedbapp/movies/presentation/SearchViewState.kt
@@ -1,0 +1,3 @@
+package com.klewerro.moviedbapp.movies.presentation
+
+data class SearchViewState(val searchText: String = "", val firstVisibleScrollIndex: Int = 0)
